Add syntax file for bible and biblereading
Bible syntax describes a bible file with the chapter in the first column and
verse in the second. Quotes are highlighted along with some other words.

Biblereading syntax describes a list of bible readings with a checkbox in the
first column for indicating which have been completed and the reading with a
book number of chapter reference. The chapter reference can have the format
"1,3,5,6" or "1-4".